# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/field/:fieldId/multiply": {
            "get": {
                "description": "Get Field Multiply",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldMultiply"
                ],
                "summary": "GetFieldMultiply",
                "operationId": "get-field-multiply",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldMultiply"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Field Multiply",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldMultiply"
                ],
                "summary": "CreateFieldMultiply",
                "operationId": "create-field-multiply",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldMultiply"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldMultiply"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Field Multiply",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldMultiply"
                ],
                "summary": "DeleteFieldMultiply",
                "operationId": "delete-field-multiply",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Field Multiply",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldMultiply"
                ],
                "summary": "UpdateFieldMultiply",
                "operationId": "update-field-multiply",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldMultiply"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldMultiply"
                        }
                    }
                }
            }
        },
...
        "/api/v1/field/:fieldId/placeholder": {
            "get": {
                "description": "Get Field Placeholder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldPlaceholder"
                ],
                "summary": "GetFieldPlaceholder",
                "operationId": "get-field-placeholder",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldPlaceholder"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Field Placeholder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldPlaceholder"
                ],
                "summary": "CreateFieldPlaceholder",
                "operationId": "create-field-placeholder",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldPlaceholder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldPlaceholder"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Field Placeholder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldPlaceholder"
                ],
                "summary": "UpdateFieldPlaceholder",
                "operationId": "update-field-placeholder",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldPlaceholder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldPlaceholder"
                        }
                    }
                }
            }
        },
        "/api/v1/field/:fieldId/range": {
            "get": {
                "description": "Get Field Range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldRange"
                ],
                "summary": "GetFieldRange",
                "operationId": "get-field-range",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldRange"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Field Range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldRange"
                ],
                "summary": "CreateFieldRange",
                "operationId": "create-field-range",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldRange"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldRange"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Field Range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldRange"
                ],
                "summary": "DeleteFieldRange",
                "operationId": "delete-field-range",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Field Range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldRange"
                ],
                "summary": "UpdateFieldRange",
                "operationId": "update-field-range",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldRange"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldRange"
                        }
                    }
                }
            }
        },
...
        "/api/v1/form/:formId/field": {
            "get": {
                "description": "Get All Fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "GetAllFields",
                "operationId": "get-all-fields",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Field"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "CreateField",
                "operationId": "create-field",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Field"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Field"
                        }
                    }
                }
            }
        },
        "/api/v1/form/:formId/field/:id": {
            "get": {
                "description": "Get One Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "GetOneField",
                "operationId": "get-one-field",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Field"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "DeleteField",
                "operationId": "delete-field",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "UpdateField",
                "operationId": "update-field",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Field"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Field"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.Field": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_of": {
                    "type": "integer"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "request.FieldMultiply": {
            "type": "object",
            "properties": {
                "is_multiply": {
                    "type": "boolean"
                }
            }
        },
        "request.FieldPlaceholder": {
            "type": "object",
            "properties": {
                "placeholder": {
                    "type": "string"
                }
            }
        },
        "request.FieldRange": {
            "type": "object",
            "properties": {
                "max": {
                    "type": "integer"
                },
                "min": {
                    "type": "integer"
                }
            }
        },
...
        "response.Field": {
            "type": "object",
            "properties": {
                "form_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_of": {
                    "type": "integer"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "response.FieldMultiply": {
            "type": "object",
            "properties": {
                "field_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_multiply": {
                    "type": "boolean"
                }
            }
        },
        "response.FieldPlaceholder": {
            "type": "object",
            "properties": {
                "field_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "placeholder": {
                    "type": "string"
                }
            }
        },
        "response.FieldRange": {
            "type": "object",
            "properties": {
                "field_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "max": {
                    "type": "integer"
                },
                "min": {
                    "type": "integer"
                }
            }
        },
...
    }
}